Finally decided to use Version Control
It has been a long way, but I finally wrap my head around using version control to track the project progress and benefit all the sweat goodness that git provides for version tracking. Really, I should done it a long time ago, even if it stayed locally. However, the main goal is to eventually publish your work online on github or gitlab. I think I can become more consistent with this. Project wise, the project is playable but not in a releasable stake. Only player movement and shoot work as intended, for the rest of the features, additional effort are mandory.