docs & fix: Write README.MD and solve shoot bug
The first intention of this commit was to build the README.MD before moving on. However, while I was taking video recording the game to extract a good screenshoot, a bug was undermining the experience and possibly the quality of the picture. Thus, I first fixed the bug where bullet were auto-deleting once in the scene, and then made my screencast. After extracting the needed screenshoot, I completed the README.MD without major issue.