Add Camera feature, Spawn Feature, and Wall
First there is still a bug on the pause menu. Since I zoon the camera, now the pause menu appear tiny on screen. Second, although the features are complete for the spawnpoint, the game is still not well balanced on that regards. Finally, the next step is probably VFX, SFX, and Game Juice . Note : This commit is quite big. I was so focused on finishing some features that it pilled up. Anyways.